+++ src/main/java/net/thesquire/backroomsmod/BackroomsMod.java
@@ -2,10 +2,13 @@ package net.thesquire.backroomsmod;
 
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
+import net.fabricmc.fabric.api.event.lifecycle.v1.ServerTickEvents;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.util.Identifier;
 import net.thesquire.backroomsmod.block.ModBlockEntities;
 import net.thesquire.backroomsmod.block.ModBlocks;
+import net.thesquire.backroomsmod.config.ModConfig;
+import net.thesquire.backroomsmod.dimension.Level11Weather;
 import net.thesquire.backroomsmod.dimension.ModDimensionKeys;
 import net.thesquire.backroomsmod.event.ModGameEvents;
 import net.thesquire.backroomsmod.item.ModItemGroup;
@@ -22,6 +25,7 @@ import net.thesquire.backroomsmod.world.feature.placement.ModPlacementModifierTy
 import net.thesquire.backroomsmod.world.gen.ModDensityFunctions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import reborncore.common.config.Configuration;
 
 //TODO fix level 2 portal generating close to level 1 portal destination
 
@@ -41,6 +45,8 @@ public class BackroomsMod implements ModInitializer {
 		// However, some things (like resources) may still be uninitialized.
 		// Proceed with mild caution.
 
+		new Configuration(ModConfig.class, BackroomsMod.MOD_ID);
+
 		ServerLifecycleEvents.SERVER_STARTED.register(server -> {
 			ServerWorld serverWorld = server.getWorld(ModDimensionKeys.LEVEL_0);
 			if(serverWorld == null) LOGGER.error("Failed to initialize level 0 portal storage", new NullPointerException());
@@ -48,6 +54,9 @@ public class BackroomsMod implements ModInitializer {
 			portalStorage.markDirty();
 		});
 
+		ServerTickEvents.START_WORLD_TICK.register(Level11Weather::handleWeather);
+
+		ModDimensionKeys.registerDimensionKeys();
 		ModServerboundPackets.registerServerboundPackets();
 		ModDensityFunctions.registerModDensityFunctions();
 		ModPlacementModifierTypes.registerPlacementModifierTypes();